refactor(datatable): move cell helpers to cells/ (refs #298)
