+++ src/datatable/cells/createdCellHelpers.test.ts
@@ -1,6 +1,6 @@
 import { ProcessStringValueStyle } from './createdCellHelpers';
 import { ColumnStyleItemType } from 'components/options/columnstyles/types';
-import { FormattedColumnValue } from './types';
+import { FormattedColumnValue } from '../types';
 import { TimeRange, dateTime } from '@grafana/data';
 
 // Covers the plumbing that forwards `replaceVariables` from BuildColumnDefs
